+++ internal/handlers/pg/pgdb/pool.go
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package pgdb provides PostgreSQL connection utilities.
 package pgdb
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"regexp"
 	"strings"
 
 	"github.com/jackc/pgconn"
@@ -29,7 +27,6 @@ import (
 	"github.com/jackc/pgx/v4/log/zapadapter"
 	"github.com/jackc/pgx/v4/pgxpool"
 	"go.uber.org/zap"
-	"golang.org/x/exp/slices"
 
 	"github.com/FerretDB/FerretDB/internal/fjson"
 	"github.com/FerretDB/FerretDB/internal/types"
@@ -44,26 +41,6 @@ const (
 	// Supported locales: (For more info see: https://www.gnu.org/software/libc/manual/html_node/Standard-Locales.html)
 	localeC     = "C"
 	localePOSIX = "POSIX"
-)
-
-// Regex validateCollectionNameRe validates collection names.
-var validateCollectionNameRe = regexp.MustCompile("^[a-zA-Z_][a-zA-Z0-9_]{0,119}$")
-
-// Errors are wrapped with lazyerrors.Error,
-// so the caller needs to use errors.Is to check the error,
-// for example, errors.Is(err, ErrSchemaNotExist).
-var (
-	// ErrTableNotExist indicates that there is no such table.
-	ErrTableNotExist = fmt.Errorf("table does not exist")
-
-	// ErrSchemaNotExist indicates that there is no such schema.
-	ErrSchemaNotExist = fmt.Errorf("schema does not exist")
-
-	// ErrAlreadyExist indicates that a schema or table already exists.
-	ErrAlreadyExist = fmt.Errorf("schema or table already exist")
-
-	// ErrInvalidTableName indicates that a schema or table didn't passed name checks.
-	ErrInvalidTableName = fmt.Errorf("invalid table name")
 )
 
 // Pool represents PostgreSQL concurrency-safe connection pool.
@@ -200,66 +177,11 @@ func (pgPool *Pool) Schemas(ctx context.Context) ([]string, error) {
 	return Databases(ctx, pgPool)
 }
 
-// Databases returns a sorted list of FerretDB database names / PostgreSQL schema names.
-func Databases(ctx context.Context, querier pgxtype.Querier) ([]string, error) {
-	sql := "SELECT schema_name FROM information_schema.schemata ORDER BY schema_name"
-	rows, err := querier.Query(ctx, sql)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-	defer rows.Close()
-
-	res := make([]string, 0, 2)
-	for rows.Next() {
-		var name string
-		if err = rows.Scan(&name); err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		if strings.HasPrefix(name, "pg_") || name == "information_schema" {
-			continue
-		}
-
-		res = append(res, name)
-	}
-	if err = rows.Err(); err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return res, nil
-}
-
 // Collections method should not be used in new code.
 //
 // Deprecated: use Collections function instead.
 func (pgPool *Pool) Collections(ctx context.Context, db string) ([]string, error) {
 	return Collections(ctx, pgPool, db)
-}
-
-// Collections returns a sorted list of FerretDB collection names.
-func Collections(ctx context.Context, querier pgxtype.Querier, db string) ([]string, error) {
-	schemaExists, err := schemaExists(ctx, querier, db)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	if !schemaExists {
-		return nil, ErrSchemaNotExist
-	}
-
-	settings, err := getSettingsTable(ctx, querier, db)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	collectionsDoc := must.NotFail(settings.Get("collections"))
-
-	collections, ok := collectionsDoc.(*types.Document)
-	if !ok {
-		return nil, lazyerrors.Errorf("invalid settings document: %v", collectionsDoc)
-	}
-
-	return collections.Keys(), nil
 }
 
 // Tables method should not be used in new code.
@@ -269,27 +191,6 @@ func (pgPool *Pool) Tables(ctx context.Context, schema string) ([]string, error)
 	return Tables(ctx, pgPool, schema)
 }
 
-// Tables returns a sorted list of PostgreSQL table names.
-// Returns empty slice if schema does not exist.
-// Tables with prefix "_ferretdb_" are filtered out.
-func Tables(ctx context.Context, querier pgxtype.Querier, schema string) ([]string, error) {
-	tables, err := tables(ctx, querier, schema)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	filtered := make([]string, 0, len(tables))
-	for _, table := range tables {
-		if strings.HasPrefix(table, reservedCollectionPrefix) {
-			continue
-		}
-
-		filtered = append(filtered, table)
-	}
-
-	return filtered, nil
-}
-
 // CreateDatabase method should not be used in new code.
 //
 // Deprecated: use CreateDatabase function instead.
@@ -297,73 +198,11 @@ func (pgPool *Pool) CreateDatabase(ctx context.Context, db string) error {
 	return CreateDatabase(ctx, pgPool, db)
 }
 
-// CreateDatabase creates a new FerretDB database (PostgreSQL schema).
-//
-// It returns (possibly wrapped) ErrAlreadyExist if schema already exist,
-// use errors.Is to check the error.
-func CreateDatabase(ctx context.Context, querier pgxtype.Querier, db string) error {
-	sql := `CREATE SCHEMA ` + pgx.Identifier{db}.Sanitize()
-	_, err := querier.Exec(ctx, sql)
-
-	if err == nil {
-		err = createSettingsTable(ctx, querier, db)
-	}
-
-	if err == nil {
-		return nil
-	}
-
-	var pgErr *pgconn.PgError
-	if !errors.As(err, &pgErr) {
-		return lazyerrors.Error(err)
-	}
-
-	switch pgErr.Code {
-	case pgerrcode.DuplicateSchema:
-		return ErrAlreadyExist
-	case pgerrcode.UniqueViolation, pgerrcode.DuplicateObject:
-		// https://www.postgresql.org/message-id/CA+TgmoZAdYVtwBfp1FL2sMZbiHCWT4UPrzRLNnX1Nb30Ku3-gg@mail.gmail.com
-		// The same thing for schemas. Reproducible by integration tests.
-		return ErrAlreadyExist
-	default:
-		return lazyerrors.Error(err)
-	}
-}
-
-// CreateDatabaseIfNotExists creates a new FerretDB database (PostgreSQL schema).
-// If the schema already exists, no error is returned.
-func CreateDatabaseIfNotExists(ctx context.Context, querier pgxtype.Querier, db string) error {
-	sql := `CREATE SCHEMA IF NOT EXISTS ` + pgx.Identifier{db}.Sanitize()
-	_, err := querier.Exec(ctx, sql)
-
-	if err == nil {
-		err = createSettingsTable(ctx, querier, db)
-	}
-
-	if err == nil || errors.Is(err, ErrAlreadyExist) {
-		return nil
-	}
-
-	var pgErr *pgconn.PgError
-	if !errors.As(err, &pgErr) {
-		return lazyerrors.Error(err)
-	}
-
-	switch pgErr.Code {
-	case pgerrcode.DuplicateSchema:
-		return ErrAlreadyExist
-	case pgerrcode.UniqueViolation, pgerrcode.DuplicateObject:
-		// https://www.postgresql.org/message-id/CA+TgmoZAdYVtwBfp1FL2sMZbiHCWT4UPrzRLNnX1Nb30Ku3-gg@mail.gmail.com
-		// The same thing for schemas. Reproducible by integration tests.
-		return ErrAlreadyExist
-	default:
-		return lazyerrors.Error(err)
-	}
-}
-
 // DropDatabase drops FerretDB database.
 //
 // It returns ErrTableNotExist if schema does not exist.
+//
+// TODO move to function
 func (pgPool *Pool) DropDatabase(ctx context.Context, db string) error {
 	sql := `DROP SCHEMA ` + pgx.Identifier{db}.Sanitize() + ` CASCADE`
 	_, err := pgPool.Exec(ctx, sql)
@@ -384,129 +223,11 @@ func (pgPool *Pool) DropDatabase(ctx context.Context, db string) error {
 	}
 }
 
-// CreateCollection creates a new FerretDB collection in existing schema.
-//
-// It returns a possibly wrapped error:
-//  * ErrInvalidTableName - if a FerretDB collection name doesn't conform to restrictions.
-//  * ErrAlreadyExist - if a FerretDB collection with the given names already exists.
-//  * ErrTableNotExist - is the required FerretDB database does not exist.
-// Please use errors.Is to check the error.
-func CreateCollection(ctx context.Context, querier pgxtype.Querier, db, collection string) error {
-	if !validateCollectionNameRe.MatchString(collection) {
-		return ErrInvalidTableName
-	}
-
-	if strings.HasPrefix(collection, reservedCollectionPrefix) {
-		return ErrInvalidTableName
-	}
-
-	schemaExists, err := schemaExists(ctx, querier, db)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	if !schemaExists {
-		return ErrSchemaNotExist
-	}
-
-	table := formatCollectionName(collection)
-	tables, err := tables(ctx, querier, db)
-	if err != nil {
-		return err
-	}
-	if slices.Contains(tables, table) {
-		return ErrAlreadyExist
-	}
-
-	settings, err := getSettingsTable(ctx, querier, db)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	collectionsDoc := must.NotFail(settings.Get("collections"))
-	collections, ok := collectionsDoc.(*types.Document)
-	if !ok {
-		return lazyerrors.Errorf("expected document but got %[1]T: %[1]v", collectionsDoc)
-	}
-
-	if collections.Has(collection) {
-		return nil
-	}
-
-	must.NoError(collections.Set(collection, table))
-	must.NoError(settings.Set("collections", collections))
-
-	err = updateSettingsTable(ctx, querier, db, settings)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	sql := `CREATE TABLE IF NOT EXISTS ` + pgx.Identifier{db, table}.Sanitize() + ` (_jsonb jsonb)`
-	if _, err = querier.Exec(ctx, sql); err == nil {
-		return nil
-	}
-
-	var pgErr *pgconn.PgError
-	if !errors.As(err, &pgErr) {
-		return lazyerrors.Error(err)
-	}
-
-	switch pgErr.Code {
-	case pgerrcode.UniqueViolation, pgerrcode.DuplicateObject:
-		// https://www.postgresql.org/message-id/CA+TgmoZAdYVtwBfp1FL2sMZbiHCWT4UPrzRLNnX1Nb30Ku3-gg@mail.gmail.com
-		// Reproducible by integration tests.
-		return ErrAlreadyExist
-	default:
-		return lazyerrors.Error(err)
-	}
-}
-
 // DropCollection method should not be used in new code.
 //
 // Deprecated: use DropCollection function instead.
 func (pgPool *Pool) DropCollection(ctx context.Context, schema, collection string) error {
 	return DropCollection(ctx, pgPool, schema, collection)
-}
-
-// DropCollection drops FerretDB collection.
-//
-// It returns (possibly wrapped) ErrTableNotExist if schema or table does not exist.
-// Please use errors.Is to check the error.
-func DropCollection(ctx context.Context, querier pgxtype.Querier, schema, collection string) error {
-	schemaExists, err := schemaExists(ctx, querier, schema)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	if !schemaExists {
-		return ErrSchemaNotExist
-	}
-
-	table := formatCollectionName(collection)
-	tables, err := tables(ctx, querier, schema)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-	if !slices.Contains(tables, table) {
-		return ErrTableNotExist
-	}
-
-	err = removeTableFromSettings(ctx, querier, schema, collection)
-	if err != nil && !errors.Is(err, ErrTableNotExist) {
-		return lazyerrors.Error(err)
-	}
-	if errors.Is(err, ErrTableNotExist) {
-		return ErrTableNotExist
-	}
-
-	// TODO https://github.com/FerretDB/FerretDB/issues/811
-	sql := `DROP TABLE IF EXISTS` + pgx.Identifier{schema, table}.Sanitize() + `CASCADE`
-	_, err = querier.Exec(ctx, sql)
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	return nil
 }
 
 // CreateTableIfNotExist ensures that given FerretDB database / PostgreSQL schema
@@ -546,19 +267,6 @@ func (pgPool *Pool) CreateTableIfNotExist(ctx context.Context, db, collection st
 // Deprecated: use CollectionExists function instead.
 func (pgPool *Pool) CollectionExists(ctx context.Context, db, collection string) (bool, error) {
 	return CollectionExists(ctx, pgPool, db, collection)
-}
-
-// CollectionExists returns true if FerretDB collection exists.
-func CollectionExists(ctx context.Context, querier pgxtype.Querier, db, collection string) (bool, error) {
-	collections, err := Collections(ctx, querier, db)
-	if err != nil {
-		if errors.Is(err, ErrSchemaNotExist) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	return slices.Contains(collections, collection), nil
 }
 
 // SchemaStats returns a set of statistics for FerretDB database / PostgreSQL schema and table.
@@ -652,9 +360,7 @@ func (pgPool *Pool) DeleteDocumentsByID(ctx context.Context, db, collection stri
 
 // InsertDocument inserts a document into FerretDB database and collection.
 // If database or collection does not exist, it will be created.
-func InsertDocument(
-	ctx context.Context, querier pgxtype.Querier, db, collection string, doc *types.Document,
-) error {
+func InsertDocument(ctx context.Context, querier pgxtype.Querier, db, collection string, doc *types.Document) error {
 	exists, err := CollectionExists(ctx, querier, db, collection)
 	if err != nil {
 		return err
@@ -680,55 +386,6 @@ func InsertDocument(
 
 	_, err = querier.Exec(ctx, sql, must.NotFail(fjson.Marshal(doc)))
 	return err
-}
-
-// tables returns a list of PostgreSQL table names.
-func tables(ctx context.Context, querier pgxtype.Querier, schema string) ([]string, error) {
-	sql := `SELECT table_name ` +
-		`FROM information_schema.columns ` +
-		`WHERE table_schema = $1 ` +
-		`GROUP BY table_name ` +
-		`ORDER BY table_name`
-	rows, err := querier.Query(ctx, sql, schema)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-	defer rows.Close()
-
-	tables := make([]string, 0, 2)
-	var name string
-	for rows.Next() {
-		if err = rows.Scan(&name); err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		tables = append(tables, name)
-	}
-	if err = rows.Err(); err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return tables, nil
-}
-
-// schemaExists returns true if given schema exists.
-func schemaExists(ctx context.Context, querier pgxtype.Querier, db string) (bool, error) {
-	sql := `SELECT nspname FROM pg_catalog.pg_namespace WHERE nspname = $1`
-	rows, err := querier.Query(ctx, sql, db)
-	if err != nil {
-		return false, lazyerrors.Error(err)
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var name string
-		must.NoError(rows.Scan(&name))
-		if name == db {
-			return true, nil
-		}
-	}
-
-	return false, nil
 }
 
 // InTransaction wraps the given function f in a transaction.
